requirements.txt
fastapi
uvicorn
motor
google-genai
python-dotenv
pydantic